requirements.txt
psutil
pywin32
Pillow
pystray